fdtget: Support decoding phandles
Currently nodes with phandles just print out as plain integer values, e.g.

   $ fdtget firefly-rk3288/u-boot.dtb /gpio-keys/button@0 gpios
   112 5 1

A common use of phandles is:

   <phandle> <n cells of stuff> <phandle> <m cells of stuff>

It is useful to be able to decode these phandles and print them out in a
properly formatted way, e.g.

   $ fdtget firefly-rk3288/u-boot.dtb -P -c gpio /gpio-keys/button@0 gpios
   /pinctrl/gpio0@ff750000 5 1

Add a -P option to decode these sorts of phandles. For those which require
arguments, add a -c option to specify the name of the '#xxx-cells' property
in the target node. This allows the tool to look up the number of cells
that follow each phandle.

This feature covers a common use of phandles. Of course it is not possible
to detect phandles automatically and other uses of phandles will not be
correctly displayed by this feature.

Signed-off-by: Simon Glass <sjg@chromium.org>
Series-to: dtc
Series-cc: david
Series-cc: Jason Clinton <jclinton@chromium.org>
Series-version: 3
Series-changes: 2
- Use 'decode' rather than 'follow'
- Be more specific in the commit message about what is and isn't supported
- Drop the unnecessary phandle_start variable

Series-changes: 3
- Change flag from -f to -P
- Change long flag from --phandle to --decode-phandles
- Update tests so that we use the --decode-phandles option as well as -P
- Add a new test for properties that don't contain phandles