RFC: Python-based device-tree validation
This is an attempt to illustrate an alternative method of validating
device-tree schema, based on pylibfdt and Python-based schema files.

Two tools are included:

   validate_dts.py  - validates a .dts file based on schema files it
			finds in the kernel
   binding_to_py.py - create Python schema files from an individual
			.txt binding file

Both tools are just at the proof-of-concept stage. See the instructions
at the top of each for usage.

This DT validator came out of work in Chrome OS and was proven successful
in providing a simple validator which is extremely flexible (e.g. custom
code is easily added to check phandle targets and other non-trivial rules)
and easy to use. It provides simple error messages when it finds a
problem, and works directly from the compiled .dtb file. It is also quite
fast even in the face of large schemas and large DT files.

Schema is handled by Python files in Documentation/devicetree/binding,
mirroring the existing .txt files. The above tool makes an attempt to
convert .txt to .py, but it is very basic at present.

To try this you also need the kernel patch:

   RFC: Example schema files written in Python

(or check README.kernel_validation

Signed-off-by: Simon Glass <sjg@chromium.org>