Add tests for fdtgrep
Add some tests to cover the current functionality.

Series-changes: 4
- Chop back to the minimal useful functionality
- Update cover letter for developments over the past 6 years
- Rebase to master

Series-to: dtc
Series-version: 4
Series-process-log: sort, uniq
Series-cc: david
Cover-letter:
Introduce fdtgrep for subsetting and hashing FDTs
Note: This was last sent 6 years ago. It really belongs upstream and I
believe it is useful functionality for libfdt, so I am trying again.
Please take a fresh look at this. It is cut back from the previous series.
If accepted we can expand the feature set piece by piece from here.

This series adds two new functions, fdt_first_region() and
fdt_next_region() which map FDT parts such as nodes and properties to
their regions in the FDT binary. The function is then used to implement
a grep utility for FDTs.

The core code is quite simple and small, but it grew a little due
to the need to make it iterative (first/next/next). Also this series adds
tests and a grep utility, so quite a bit of code is built on it.

The use for this feature is twofold. Firstly it provides a convenient
way of performing a structure-aware grep of the tree. For example it is
possible to grep for a node and get all the properties associated with
that node. Trees can be subsetted easily, by specifying the nodes that
are required, and then writing out the regions returned by this function.
This is useful for small resource-constrained systems, such as boot
loaders, which want to use an FDT but do not need to know about all of
it. The full FDT can be grepped to pull out the few things that are
needed - this can be an automatic part of the build system and does
not require the FDT source code.

This first use makes it easy to implement an FDT grep tool. Options are
provided to search for named nodes and properties. It is possible to
search for non-matches also (useful for excluding a particular property
from the FDT, for example). The output is like fdtdump, but only with
the regions selected by the grep. The fdtgrep utility cannot currently
output valid source in all cases, which can be used by dtc. That will
come later.

Secondly it makes it easy to hash parts of the tree and detect changes.
The intent is to get a list of regions which will be invariant provided
those parts are invariant. For example, if you request a list of regions
for all nodes but exclude the property "data", then you will get the
same region contents regardless of any change to "data" properties.
An assumption is made here that the tree ordering remains the same.

This second use is the mechanism behind U-Boot's verified-boot feature
using Flat Image Tree (FIT) images, introduced 6 years ago. Briefly, this
works by signing configurations (consisting of particular kernel and FDT
combinations) so that the boot loader can verify that these combinations
are valid and permitted. Since a FIT image is in fact an FDT, we need to
be able to hash particular regions of the FDT for the signing and
verification process. This is done by using the region functions to
select the data that needs to be hashed for a particular configuration.

A third use is currently under discussion in the boot-architecture
mailing list, which is to sign vanilla FDT files by adding a signature
node. This will avoid the need for a separate signature for every device
tree file, or inventing yet another file format.

The fdtgrep utility could be used to replace all of the functions of
fdtdump. However fdtdump is intended as a separate, simple way of
dumping the tree (for verifying against dtc output for example). So
fdtdump remains a separate program and this series leaves it alone.

Note: a somewhat unfortunately feature of this implementation is that
a state structure needs to be kept around between calls of
fdt_next_region(). This is declared in libfdt.h but really should be
opaque. Ideas welcome.
END

Signed-off-by: Simon Glass <sjg@chromium.org>