Add fdtgrep to grep and hash FDTs
fdtgrep operates like a grep for device trees, working on binary .dtb
files.

fdtgrep can find nodes by name or compatible string. It can find
properties by name. Using a list of nodes/properties to match, fdtgrep
creates either .dts text output containing just those matches, or binary
output.

The raw binary output, without the normal FDT headers, can be useful for
hashing part of an FDT and making sure that that part does not change,
even if other parts do. This makes it possible to detect an important
change, while ignoring a change that is of no interest.

At its simplest fdtgrep can be used as follows;

   fdtgrep /holiday tests/grep.dtb

and it produces just that node (with a skeleton to hold it):

/ {
    holiday {
        compatible = "ixtapa", "mexico";
        weather = "sunny";
        status = "okay";
    };
};

The binary form of this is:

   fdtgrep -n /holiday -O dtb tests/grep.dtb

which produces a fragment of the dtb with just the above node.

Various options are provided to search for properties and to ensure that
nodes/properties are not present.

Series-changes: 4
- Chop back to the minimal useful functionality

Series-changes: 3
- Adjust help and command line processing to follow new approach
- Rename -V to -I to avoid using -V for a different purpose to other tools
- Rename -s to -e since it only 'enters' the node and does not include it all
- Add -s option to include all subnodes
- Add -f option to display offset; make -a display an absolute file address
- Add documentation on fdtgrep

Series-changes: 2
- Add local fdt_find_regions() function since libfdt no longer has it

Signed-off-by: Simon Glass <sjg@chromium.org>